requirements.txt
yt-dlp
ipywidgets
google-cloud-storage
google-cloud-secret-manager